+++ packages/api/endpoints/collections.js
@@ -4,6 +4,7 @@ const { handle } = require('../lib/response');
 const models = require('../models');
 const Collection = require('../es/collections');
 const RecordDoesNotExist = require('../lib/errors').RecordDoesNotExist;
+const { injectConceptId } = require('../lib/injectConceptId');
 
 /**
  * List all collections.
@@ -14,7 +15,10 @@ const RecordDoesNotExist = require('../lib/errors').RecordDoesNotExist;
  */
 function list(event, cb) {
   const collection = new Collection(event);
-  return collection.query().then((res) => cb(null, res)).catch(cb);
+  return collection.query()
+    .then((res) => injectConceptId(res))
+    .then((res) => cb(null, res))
+    .catch(cb);
 }
 
 /**